+++ src/client/modules/Tasks/TaskDetails/TaskButtons.jsx
@@ -1,12 +1,16 @@
 import React from 'react'
+import Link from '@govuk-react/link'
+import Button from '@govuk-react/button'
+import { connect } from 'react-redux'
 
 import { Form, FormLayout } from '../../../components'
 
 import urls from '../../../../lib/urls'
-import { TASK_ARCHIVE_TASK } from './state'
+import { TASK_ARCHIVE_TASK, state2props } from './state'
 import { FORM_LAYOUT } from '../../../../common/constants'
+import { GREY_3, TEXT_COLOUR } from '../../../utils/colours'
 
-const TaskButtons = ({ task }) => (
+const TaskButtons = ({ task, editUrl }) => (
   <>
     {!task.archived && (
       <FormLayout setWidth={FORM_LAYOUT.THREE_QUARTERS}>
@@ -31,7 +35,15 @@ const TaskButtons = ({ task }) => (
         </Form>
       </FormLayout>
     )}
+    <Button
+      buttonColour={GREY_3}
+      buttonTextColour={TEXT_COLOUR}
+      as={Link}
+      href={editUrl}
+    >
+      Edit
+    </Button>
   </>
 )
 
-export default TaskButtons
+export default connect(state2props)(TaskButtons)
