+++ src/client/modules/Tasks/TaskDetails/index.jsx
@@ -9,6 +9,7 @@ import { ID, TASK_GET_TASK_DETAILS, state2props } from './state'
 import { TASK_DETAILS_LOADED } from '../../../actions'
 import Task from '../../../components/Task'
 
+//TODO move this into the state2props
 const getCompany = (task) => {
   return task?.investmentProjectTask?.investmentProject?.investorCompany
 }
