+++ src/client/modules/Tasks/TaskDetails/state.js
@@ -1,6 +1,21 @@
+import { get } from 'lodash'
+
+import urls from '../../../../lib/urls'
+
 export const ID = 'taskDetails'
 
 export const TASK_GET_TASK_DETAILS = 'TASK_GET_TASK_DETAILS'
 export const TASK_ARCHIVE_TASK = 'TASK_ARCHIVE_TASK'
 
-export const state2props = (state) => ({ ...state[ID] })
+const getEditUrl = ({ task }) => {
+  const investmentProject = get(task, 'investmentProjectTask.investmentProject')
+  return investmentProject
+    ? urls.investments.projects.tasks.edit(investmentProject.id, task.id)
+    : undefined
+}
+
+export const state2props = (state) => {
+  const task = state[ID]
+
+  return { ...task, editUrl: getEditUrl(task) }
+}
