+++ src/client/modules/Tasks/TaskForm/state.js
@@ -1,0 +1,1 @@
+export const TASK_SAVE_TASK_DETAILS = 'TASK_SAVE_TASK_DETAILS'
