+++ src/client/modules/Tasks/TaskForm/tasks.js
@@ -1,0 +1,9 @@
+import { apiProxyAxios } from '../../../components/Task/utils'
+import { transformTaskFormValuesForAPI } from './transformers'
+
+export const saveTaskDetail = ({ values, currentAdviserId, taskId }) => {
+  return apiProxyAxios.patch(
+    `/v4/task/${taskId}`,
+    transformTaskFormValuesForAPI(values, currentAdviserId)
+  )
+}
